+++ Changelog.md
@@ -6,6 +6,7 @@
 - fix incorrect advancement format
 - fix crashes on fabric when wearing the ancient helmet
 - fix monkey teleporting to player, even though they were ordered to sit
+- elephant inventory can now be opened while riding (pressing the `open-inventory`-key)
 
 ### Version 1.18.2 - 2.0.3